make setExpressions async
Used a callback before. There might not be any reason to have async.
The idea was that expressions can be large so might take a while to
compile and so maybe we could move them off the main thread. For example
they could be checked for errors in a worker, if they fail just pass
back the error, if they succeed then pass them to the
AudioWorkletProcess.