use return 0,<exp> to try to deal with comments.
The problem is JavaScript's idiotic automatic semicolon insertion.

If you have an expression like

```js
    return
     2
```

That's just returns undefined. So, if the user enters

```
// hello
t
```

That gets converted to

```js
return // hello
t
```

We used to try to strip linefeeds and comments and it mostly worked
but the comment stripper was simple because who the F wants to write
a whole effing parser just to strip comments.

As a try, if we change the generated code to

```js
return 0,// hello
t
```

The automatic semicolon insertion doesn't happen and it correctly
returns t.